实现动态更换用户信息

+++ app/src/main/java/com/example/classroom_location/Login.java
@@ -3,7 +3,6 @@ package com.example.classroom_location;
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.graphics.Color;
 import android.preference.PreferenceManager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -15,13 +14,14 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 @SuppressLint("Registered")
-public class Login extends AppCompatActivity implements View.OnClickListener,
-        View.OnFocusChangeListener {
+public class Login extends AppCompatActivity implements View.OnClickListener{
 
     private CheckBox RememberPass;
     private EditText account_edit;
     private EditText password_edit;
     private SharedPreferences preferences;
+
+//    private final String TAG = "LoginActivity";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -46,34 +46,30 @@ public class Login extends AppCompatActivity implements View.OnClickListener,
     public void onClick(View v) {
         String mAccount = "2016010599";
         String mPassword = "123456";
-        switch (v.getId()){
-            case R.id.Login_btn:
-                String account = account_edit.getText().toString();
-                String password = password_edit.getText().toString();
-                if ( account.equals(mAccount) && password.equals(mPassword) ){
-                    SharedPreferences.Editor editor = preferences.edit();
-                    if ( RememberPass.isChecked() ){    //检查复选框是否被选中
-                        editor.putBoolean("remember_password", true);
-                        editor.putString("account", account);
-                        editor.putString("password", password);
-                    }
-                    else {
-                        editor.clear();
-                    }
-                    editor.apply();
-                    Intent intent = new Intent(Login.this, MainActivity.class);
-                    startActivity(intent);
-                }
-                else{
-                    Toast.makeText(Login.this, "账号或密码错误",
-                            Toast.LENGTH_SHORT).show();
-                }
-                break;
-            default:
-                break;
-        }
+        if (v.getId() == R.id.Login_btn) {
+            String account = account_edit.getText().toString();
+            String password = password_edit.getText().toString();
+            if (account.equals(mAccount) && password.equals(mPassword)) {
+                SharedPreferences.Editor editor = preferences.edit();
+                if (RememberPass.isChecked()) {    //检查复选框是否被选中
+                    editor.putBoolean("remember_password", true);
+                    editor.putString("account", account);
+                    editor.putString("password", password);
 
+                } else {
+                    editor.clear();
+                }
+                editor.apply();
+                Intent intent = new Intent(Login.this, MainActivity.class);
+                startActivity(intent);
+            } else {
+                Toast.makeText(Login.this, "账号或密码错误",
+                        Toast.LENGTH_SHORT).show();
+            }
+        }
     }
+
+
 
     private void init(){
         //  将账号和密码都设置到文本框中
@@ -84,22 +80,5 @@ public class Login extends AppCompatActivity implements View.OnClickListener,
         RememberPass.setChecked(true);
     }
 
-    @Override
-    public void onFocusChange(View v, boolean hasFocus) {
-        switch (v.getId()){
-            case R.id.account_edit:
-                if (hasFocus){
-                    account_edit.setHintTextColor(Color.parseColor("#FFFFFF"));
-                }
-                else {
-                    account_edit.setHintTextColor(Color.parseColor("#515151"));
-                }
-                break;
-            case R.id.password_edit:
 
-                break;
-            default:
-                break;
-        }
-    }
 }
